SpectrumTransformer in lib-fft and TrackSpectrumTransformer in lib-wave-track-fft
