+++ libraries/lib-fft/SpectrumTransformer.cpp
@@ -12,7 +12,6 @@ Edward Hui
 
 #include <algorithm>
 #include "FFT.h"
-#include "WaveTrack.h"
 
 SpectrumTransformer::SpectrumTransformer( bool needsOutput,
    eWindowFunctions inWindowType,
@@ -98,12 +97,6 @@ bool SpectrumTransformer::DoStart()
 bool SpectrumTransformer::DoFinish()
 {
    return true;
-}
-
-void
-TrackSpectrumTransformer::DoOutput(const float *outBuffer, size_t mStepSize)
-{
-   mOutputTrack->Append((constSamplePtr)outBuffer, floatSample, mStepSize);
 }
 
 bool SpectrumTransformer::Start(size_t queueLength)
@@ -338,61 +331,6 @@ bool SpectrumTransformer::QueueIsFull() const
       return (mOutStepCount >= 0);
 }
 
-bool TrackSpectrumTransformer::Process(const WindowProcessor &processor,
-   const WaveChannel &channel, size_t queueLength, sampleCount start,
-   sampleCount len)
-{
-   mpChannel = &channel;
-
-   if (!Start(queueLength))
-      return false;
-
-   auto bufferSize = channel.GetMaxBlockSize();
-   FloatVector buffer(bufferSize);
-
-   bool bLoopSuccess = true;
-   auto samplePos = start;
-   while (bLoopSuccess && samplePos < start + len) {
-      //Get a blockSize of samples (smaller than the size of the buffer)
-      const auto blockSize = limitSampleBufferSize(
-         std::min(bufferSize, channel.GetBestBlockSize(samplePos)),
-         start + len - samplePos);
-
-      //Get the samples from the track and put them in the buffer
-      channel.GetFloats(buffer.data(), samplePos, blockSize);
-      samplePos += blockSize;
-      bLoopSuccess = ProcessSamples(processor, buffer.data(), blockSize);
-   }
-
-   if (!Finish(processor))
-      return false;
-
-   return bLoopSuccess;
-}
-
-bool TrackSpectrumTransformer::DoFinish()
-{
-   return SpectrumTransformer::DoFinish();
-}
-
-bool TrackSpectrumTransformer::PostProcess(
-   WaveTrack &outputTrack, sampleCount len)
-{
-   outputTrack.Flush();
-   auto tLen = outputTrack.LongSamplesToTime(len);
-   // Filtering effects always end up with more data than they started with.
-   // Delete this 'tail'.
-   outputTrack.Clear(tLen, outputTrack.GetEndTime());
-   return true;
-}
-
 SpectrumTransformer::~SpectrumTransformer() = default;
 
 SpectrumTransformer::Window::~Window() = default;
-
-TrackSpectrumTransformer::~TrackSpectrumTransformer() = default;
-
-bool TrackSpectrumTransformer::DoStart()
-{
-   return SpectrumTransformer::DoStart();
-}
