+++ libraries/lib-fft/SpectrumTransformer.h
@@ -11,28 +11,26 @@ Paul Licameli
 
 #ifndef __AUDACITY_SPECTRUM_TRANSFORMER__
 #define __AUDACITY_SPECTRUM_TRANSFORMER__
- 
+
+#include <cassert>
 #include <functional>
 #include <memory>
 #include <vector>
-#include "audacity/Types.h"
 #include "RealFFTf.h"
 #include "SampleCount.h"
 
 enum eWindowFunctions : int;
 
-class WaveChannel;
-
 /*!
  @brief A class that transforms a portion of a wave track (preserving duration)
  by applying Fourier transform, then modifying coefficients, then inverse
  Fourier transform and overlap-add to reconstruct.
- 
+
  @par The procedure that modifies coefficients can be varied, and can employ lookahead
  and -behind to nearby windows.  May also be used just to gather information
  without producing output.
 */
-class SpectrumTransformer /* not final */
+class FFT_API SpectrumTransformer /* not final */
 {
 public:
    // Public interface
@@ -83,7 +81,7 @@ public:
    bool Finish(const WindowProcessor &processor);
 
    //! Derive this class to add information to the queue.  @see NewWindow()
-   struct Window
+   struct FFT_API Window
    {
       explicit Window(size_t windowSize)
          : mRealFFTs( windowSize / 2 )
@@ -157,7 +155,7 @@ protected:
 
    const unsigned mStepsPerWindow;
    const size_t mStepSize;
-   
+
    const bool mLeadingPadding;
 
    const bool mTrailingPadding;
@@ -178,46 +176,6 @@ private:
    FloatVector mOutWindow;
 
    const bool mNeedsOutput;
-};
-
-class WaveTrack;
-
-//! Subclass of SpectrumTransformer that rewrites a track
-class TrackSpectrumTransformer /* not final */ : public SpectrumTransformer {
-public:
-   /*!
-    @copydoc SpectrumTransformer::SpectrumTransformer(bool,
-       eWindowFunctions, eWindowFunctions, size_t, unsigned, bool, bool)
-    @pre `!needsOutput || pOutputTrack != nullptr`
-    */
-   TrackSpectrumTransformer(WaveChannel *pOutputTrack,
-      bool needsOutput, eWindowFunctions inWindowType,
-      eWindowFunctions outWindowType, size_t windowSize,
-      unsigned stepsPerWindow, bool leadingPadding, bool trailingPadding
-   )  : SpectrumTransformer{ needsOutput, inWindowType, outWindowType,
-         windowSize, stepsPerWindow, leadingPadding, trailingPadding
-      }
-      , mOutputTrack{ pOutputTrack }
-   {
-      assert(!needsOutput || pOutputTrack != nullptr);
-   }
-   ~TrackSpectrumTransformer() override;
-
-   //! Invokes Start(), ProcessSamples(), and Finish()
-   bool Process(const WindowProcessor &processor, const WaveChannel &channel,
-      size_t queueLength, sampleCount start, sampleCount len);
-
-   //! Final flush and trimming of tail samples
-   static bool PostProcess(WaveTrack &outputTrack, sampleCount len);
-
-protected:
-   bool DoStart() override;
-   void DoOutput(const float *outBuffer, size_t mStepSize) override;
-   bool DoFinish() override;
-
-private:
-   WaveChannel *const mOutputTrack;
-   const WaveChannel *mpChannel = nullptr;
 };
 
 #endif
